Move FraPrint logic to Print component
