--- src/client/pages/Print/FraPrint/FraPrint.scss
@@ -1,10 +1,0 @@
-.fra-print__header {
-  display: grid;
-  align-items: center;
-  grid-row-gap: 10px;
-
-  > * {
-    display: flex;
-    align-items: center;
-  }
-}
