--- src/client/pages/Print/FraPrint/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FraPrint'
