+++ src/client/pages/Print/hooks/useGetTableSections.tsx
